+++ CHANGES.rst
@@ -2,8 +2,8 @@ Changelog
 ---------
 
 
-2.x (unreleased)
-~~~~~~~~~~~~~~~~
+2.1.3 (2019-01-09)
+~~~~~~~~~~~~~~~~~~
 
 * Fix a bug where method ``with_context`` returns an error if we update
   the values of the logged-in user before.